Add `$UuidHex` builtin
