+++ src/lib/builtins/uuid-hex.ts
@@ -1,0 +1,17 @@
+import { Ucs2StringType } from "../types/ucs2-string";
+
+/**
+ * Semantic alias for strings representing UUID.
+ */
+export type UuidHex = string;
+
+/**
+ * Type for UUID hex string standard representation.
+ */
+export const $UuidHex: Ucs2StringType = new Ucs2StringType({
+  lowerCase: true,
+  trimmed: true,
+  minLength: 36,
+  maxLength: 36,
+  pattern: /[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}/,
+});
